FFBE-14 오픈베타 홈화면 작품리스트 GET

+++ src/app/models/user_rating.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
@@ -7,7 +7,7 @@ from app.db.base_class import Base
 class UserRating(Base):
     user_id = Column(Integer, ForeignKey('user.id'), primary_key=True, index=True)
     series_id = Column(Integer, ForeignKey('series.id'), primary_key=True, index=True)
-    rating = Column(Integer, nullable=False)
+    rating = Column(Float, nullable=False)
 
     # Many to Many relation
     user = relationship('User', back_populates='user_rating')
